refactor(hompage): rearrange files in appropriate folder
explain the purpose of file more explicity. files in components folder must be reuseable

resovled issue #27
